# Avoid normalisation where possible