--- a/src/full/Agda/Compiler/MAlonzo/Compiler.hs
+++ b/src/full/Agda/Compiler/MAlonzo/Compiler.hs
@@ -252,7 +252,6 @@ definition env isMain def@Defn{defName = q, defType = ty, theDef = d} = do
           else do
             -- Make sure we have imports for all names mentioned in the type.
             hsty <- haskellType q
-            ty   <- normalise ty
             sequence_ [ xqual x (HS.Ident "_") | x <- Set.toList (namesIn ty) ]
 
           -- Check that the function isn't INLINE (since that will make this