--- a/src/full/Agda/TypeChecking/Conversion.hs
+++ b/src/full/Agda/TypeChecking/Conversion.hs
@@ -1261,28 +1261,17 @@ leqSort s1 s2 = (catchConstraint (SortCmp CmpLeq s1 s2) :: m () -> m ()) $ do
     __IMPOSSIBLE__
 
 leqLevel :: MonadConversion m => Level -> Level -> m ()
-leqLevel a b = do
-  reportSDoc "tc.conv.nat" 30 $
-    "compareLevel" <+>
-      sep [ prettyTCM a <+> "=<"
-          , prettyTCM b ]
-  leqView a b
-  where
-    -- Andreas, 2016-09-28
-    -- If we have to postpone a constraint, then its simplified form!
-    leqView :: MonadConversion m => Level -> Level -> m ()
-    leqView a b = catchConstraint (LevelCmp CmpLeq a b) $ do
+leqLevel a b = catchConstraint (LevelCmp CmpLeq a b) $ do
       reportSDoc "tc.conv.level" 30 $
-        "compareLevelView" <+>
-          sep [ pretty a <+> "=<"
-              , pretty b ]
+        "compareLevel" <+>
+          sep [ prettyTCM a <+> "=<"
+              , prettyTCM b ]
 
+      (a, b) <- reduce (a, b)
       ((a, b), equal) <- SynEq.checkSyntacticEquality a b
       reportSDoc "tc.conv.level" 60 $
         "checkSyntacticEquality returns" <+> prettyTCM equal
       unless equal $ do
-
-      (a, b) <- reduce (a, b)
 
       cumulativity <- optCumulativity <$> pragmaOptions
       reportSDoc "tc.conv.level" 40 $
@@ -1314,20 +1303,20 @@ leqLevel a b = do
 
         -- ⊔ as ≤ single
         (as@(_:|_:_), b :| []) ->
-          sequence_ [ leqView (unSingleLevel a') (unSingleLevel b) | a' <- NonEmpty.toList as ]
+          sequence_ [ leqLevel (unSingleLevel a') (unSingleLevel b) | a' <- NonEmpty.toList as ]
 
         -- reduce constants
         (as, bs)
           | let minN = min (fst $ levelPlusView a) (fst $ levelPlusView b)
                 a'   = fromMaybe __IMPOSSIBLE__ $ subLevel minN a
                 b'   = fromMaybe __IMPOSSIBLE__ $ subLevel minN b
-          , minN > 0 -> leqView a' b'
+          , minN > 0 -> leqLevel a' b'
 
         -- remove subsumed
         -- Andreas, 2014-04-07: This is ok if we do not go back to equalLevel
         (as, bs)
           | (subsumed@(_:_) , as') <- List.partition isSubsumed (NonEmpty.toList as)
-          -> leqView (unSingleLevels as') b
+          -> leqLevel (unSingleLevels as') b
           where
             isSubsumed a = any (`subsumes` a) (NonEmpty.toList bs)
 
@@ -1401,12 +1390,9 @@ leqLevel a b = do
         isMetaLevel (SinglePlus (Plus _ UnreducedLevel{})) = __IMPOSSIBLE__
         isMetaLevel _ = False
 
-equalLevel :: MonadConversion m => Level -> Level -> m ()
-equalLevel a b = equalLevel' a b
-
 -- | Precondition: levels are 'normalise'd.
-equalLevel' :: forall m. MonadConversion m => Level -> Level -> m ()
-equalLevel' a b = do
+equalLevel :: forall m. MonadConversion m => Level -> Level -> m ()
+equalLevel a b = do
   reportSDoc "tc.conv.level" 50 $ sep [ "equalLevel", nest 2 $ parens $ pretty a, nest 2 $ parens $ pretty b ]
   -- Andreas, 2013-10-31 remove common terms (that don't contain metas!)
   -- THAT's actually UNSOUND when metas are instantiated, because
@@ -1425,12 +1411,11 @@ equalLevel' a b = do
                ]
         ]
 
+  (a, b) <- reduce (a, b)
   ((a, b), equal) <- SynEq.checkSyntacticEquality a b
   reportSDoc "tc.conv.level" 60 $
     "checkSyntacticEquality returns" <+> prettyTCM equal
   unless equal $ do
-
-  (a, b) <- reduce (a, b)
 
   -- Jesper, 2014-02-02 remove terms that certainly do not contribute
   -- to the maximum
@@ -1480,9 +1465,9 @@ equalLevel' a b = do
 
         -- 0 == a ⊔ b
         (SingleClosed 0 :| [] , bs@(_:|_:_)) ->
-          sequence_ [ equalLevel' (ClosedLevel 0) (unSingleLevel b') | b' <- NonEmpty.toList bs ]
+          sequence_ [ equalLevel (ClosedLevel 0) (unSingleLevel b') | b' <- NonEmpty.toList bs ]
         (as@(_:|_:_) , SingleClosed 0 :| []) ->
-          sequence_ [ equalLevel' (unSingleLevel a') (ClosedLevel 0) | a' <- NonEmpty.toList as ]
+          sequence_ [ equalLevel (unSingleLevel a') (ClosedLevel 0) | a' <- NonEmpty.toList as ]
 
         -- meta == any
         (SinglePlus (Plus k (MetaLevel x as)) :| [] , bs)