--- a/test/interaction/Issue3032.out
+++ b/test/interaction/Issue3032.out
@@ -4,5 +4,5 @@
 (agda2-status-action "")
 (agda2-info-action "*All Goals*" "?0 : B " nil)
 ((last . 1) . (agda2-goals-action '(0)))
-((last . 2) . (agda2-make-case-action '("foo _ (c x) = ?")))
+((last . 2) . (agda2-make-case-action '("foo .(foo t x) (c x) = ?")))
 ((last . 1) . (agda2-goals-action '(0)))