In https://github.com/agda/agda/pull/4531#issuecomment-603173162 @nad said:

> It might make sense to inspect every occurrence of `normalise` in the code and see if it can be removed.

I now started going over all calls to `normalise` in the Agda codebase to see which ones could be removed. This is a first batch of cases that seemed harmless to remove, more will follow later (hopefully).