@jespercockx 2020-04-05
I pushed some more patches removing (seemingly) unnecessary uses of `normalise`. After these patches, the following potentially hazardous uses of `normalise` remain:

- Cases where we normalise to get rid of a specific variable occurrence:
```
./TypeChecking/Rules/LHS/Unify.hs:1196:          u <- liftTCM $ normalise u
./TypeChecking/Rules/LHS/Unify.hs:1197:          s <- liftTCM $ lensVarTel normalise s
./TypeChecking/Positivity.hs:544:            tel' <- normalise $ telFromList $ drop np $ telToList tel
./TypeChecking/Positivity.hs:579:      getOccurrences vars =<< normalise tel' -- Andreas, 2017-01-01, issue #1899, treat like data types
```
- Cases related to `with`-abstraction:
```
./TypeChecking/Rules/Def.hs:946:        vtys0 <- normalise vtys0
./TypeChecking/Rules/Def.hs:1043:  delta1 <- normalise delta1 -- Issue 1332: checkInternal is picky about argInfo
./TypeChecking/With.hs:141:  d2b  <- normalise d2b
./TypeChecking/With.hs:146:  vtys <- etaContract =<< normalise vtys
./TypeChecking/With.hs:442:         Def d es <- liftTCM $ normalise (unEl $ unDom a)
```
- Checking builtin definitions:
```
./TypeChecking/Rules/Builtin.hs:548:  t <- etaContract =<< normalise t
```
- Unquoting stuff:
```
./TypeChecking/Unquote.hs:113:  t <- liftTCM $ etaContract =<< normalise =<< tm
```
- Auto:
```
./Auto/Auto.hs:139:        normalise =<< do TCM.defType <$> getConstInfo def
./Auto/Auto.hs:383:                 ps <- addContext tel $ normalise ps
./Auto/Auto.hs:405:       normalise targettype
./Auto/Auto.hs:422:              ctyp <- normalise $ defType c
./Auto/Auto.hs:438:          ctyp <- normalise $ defType c
./Auto/Convert.hs:108:     typ <- lift $ normalise typ
./Auto/Convert.hs:182:        targettype <- normalise targettype
./Auto/Convert.hs:183:        localVars <- mapM normalise localVars
./Auto/Convert.hs:279:  neqc <- normalise eqc
./Auto/Convert.hs:319:    body' <- traverse convert =<< lift (normalise body)
```