@WolframKahl 2020-04-05
> `+  termToPattern =<< do liftTCM $ stripAllProjections t`
> 
> Why is hint-man complaining here? Removing the `do` leads to a parse error.

To consider:
```haskell
   termToPattern =<< liftTCM (stripAllProjections t)
```