@jespercockx 2020-04-05
*Updates*:

1. It is possible that there is some performance to be gained by only normalising terms that actually contain occurrences of the variables we want to get rid of, but I'd first want to see a test case where this is actually the bottleneck.
2. I removed two cases of normalisation from with-abstraction, the other three are essential to the current specification of `with` so getting rid of them would mean a major rework of with-abstraction as a whole.
3. and 4. I managed to remove these two as well
5. I'm not touching Agsy.

I think this is in a good state to merge. I did not benchmark it, but I do not really expect any major performance gains since most calls to `normalise` were quite innocuous.