@jespercockx 2020-04-05
It seems I broke the build, but I don't understand why. `make test` runs correctly on my system.