@ajrouvoet 2020-04-06
Ab-normalisation seems to have dramatically improved type-checking performance for my project.
A `rm -r _build && agda -vprofile:7  ./Everything.agda` before and after:

*Before*
```
Total                        211,431ms            
Miscellaneous                  1,487ms            
Typing                         9,898ms (122,565ms)
Typing.InstanceSearch         36,589ms            
Typing.CheckRHS               26,675ms            
Typing.CheckLHS                7,173ms  (14,961ms)
Typing.CheckLHS.UnifyIndices   7,787ms            
Typing.OccursCheck            12,750ms            
Typing.With                   10,827ms            
Typing.TypeSig                 7,705ms            
Typing.Generalize              3,156ms            
Serialization                 17,158ms  (29,155ms)
Serialization.BinaryEncode     6,988ms            
Serialization.BuildInterface   2,956ms            
Serialization.Sort             1,176ms            
Serialization.Compress           876ms            
Deserialization               24,315ms  (28,785ms)
Deserialization.Compaction     4,470ms            
Parsing                          497ms   (8,713ms)
Parsing.OperatorsExpr          6,306ms            
Parsing.OperatorsPattern       1,909ms            
Scoping                        2,512ms   (6,279ms)
Scoping.InverseScopeLookup     3,767ms            
DeadCode                       4,445ms            
Positivity                     3,472ms            
Coverage                       2,813ms            
Termination                      431ms   (1,341ms)
Termination.Compare              588ms            
Termination.RecCheck             269ms            
Termination.Graph                 53ms            
Highlighting                     995ms            
Import                           701ms            
Injectivity                      641ms            
ProjectionLikeness                32ms
```

*After*
```
Total                        133,218ms           
Miscellaneous                  1,259ms           
Typing                         5,778ms (71,906ms)
Typing.InstanceSearch         32,912ms           
Typing.CheckRHS               11,826ms           
Typing.OccursCheck             7,243ms           
Typing.CheckLHS                3,152ms  (5,951ms)
Typing.CheckLHS.UnifyIndices   2,798ms           
Typing.With                    5,421ms           
Typing.TypeSig                 2,058ms           
Typing.Generalize                712ms           
Deserialization               21,229ms (25,090ms)
Deserialization.Compaction     3,860ms           
Serialization                 10,433ms (17,420ms)
Serialization.BinaryEncode     3,609ms           
Serialization.BuildInterface   2,483ms           
Serialization.Compress           511ms           
Serialization.Sort               382ms           
Scoping                        1,098ms  (4,383ms)
Scoping.InverseScopeLookup     3,285ms           
DeadCode                       4,274ms           
Parsing                          257ms  (3,376ms)
Parsing.OperatorsExpr          2,284ms           
Parsing.OperatorsPattern         833ms           
Positivity                     2,440ms           
Coverage                       1,329ms           
Import                           660ms           
Termination                      158ms    (548ms)
Termination.Compare              187ms           
Termination.RecCheck             174ms           
Termination.Graph                 28ms           
Highlighting                     430ms           
Injectivity                       97ms
```