@andreasabel 2020-04-15
> To consider:
> 
> ```haskell
>    termToPattern =<< liftTCM (stripAllProjections t)
> ```

Well, in this toy case this would be an alternative, but in general, if I want to use `$` after `=<<`, I simply insert a `do` after `=<<`, and thus reset the precedence, and I can use any operator without having to worry what its relative binding strength is compared to `=<<`.

Hint-man just does not support our style, why force is own mindless suggestions on us? 

We could as well place a `do` after _every_ such `=<<` to enforce uniformity?  What is the problem with "redundant do" in the first place?  Do we have to pay tax on each "do"?